Ensure focus consistency (#123)
* Overhaul focusing strategies

Updates all the code related to focusing surfaces, centralizing the handling
code. This fixes a few issues, namely:

-Views being unmapped in workspaces other than the current would set focus to a
surface that is not on the screen;
- XWayland applications would think they always had focus.

* Unify focus clearing workflow when there are no available views

Unification ensures surface activeness is kept up to date

* Remove unused parameter from viv_view_focus

Removes the unused surface parameter from viv_view_focus, since the picking of
the surface is now automattic based on the view

* Update mouse focus on workspace change

Just like with keyboard focus, update the mouse focus when the workspace
changes. Thin ensures that surfaces get information right away, rather than
having to wait for the user to move the mouse