Delete old rendering code
This is a lot of code! It used to be necessary to handle all the damage tracking.

This leaves rendering broken for non-xdg views and for xdg popups etc. as that isn't
implemented with wlr_scene yet, but since wlr_scene clearly is suitable for use it's
easier to drop the baggage before adding back in the wlr_scene version of the
functionality.